chore(app): rename to hyphen separated file name in lib/utils
