+++ lib/craftperson/story.js
@@ -1,5 +1,5 @@
 import Octokat from 'octokat';
-import { decryptData } from '../utils/cryptData';
+import { decryptData } from '../utils/crypt-data';
 import { fetchAll } from '../utils/octokat-helper';
 export const TYPE_PULL = 'TYPE_PULL';
 export const TYPE_ISSUE = 'TYPE_ISSUE';
