+++ lib/utils/configs.js
@@ -5,7 +5,7 @@ import leveljs from 'level-js';
 import {
   decryptData,
   encryptData,
-} from './cryptData';
+} from './crypt-data';
 import moment from 'moment';
 import {
   isUndefined,
