+++ lib/utils/crypt-data.js
